event-gateway/requirements.txt
confluent-kafka==2.8.0
fastapi==0.115.8
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-sdk==1.28.2
prometheus-client==0.21.1
uvicorn[standard]==0.34.0